requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
mistralai
python-dotenv
